Add support for gateway and sub devices
