+++ bundles/org.smarthomej.binding.tuya/src/main/java/org/smarthomej/binding/tuya/internal/TuyaHandlerFactory.java
@@ -14,36 +14,32 @@ package org.smarthomej.binding.tuya.internal;
 
 import static org.smarthomej.binding.tuya.internal.TuyaBindingConstants.*;
 
-import java.lang.reflect.Type;
-import java.util.List;
 import java.util.Set;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
 import org.eclipse.jetty.client.HttpClient;
 import org.openhab.core.io.net.http.HttpClientFactory;
-import org.openhab.core.storage.Storage;
 import org.openhab.core.storage.StorageService;
+import org.openhab.core.thing.Bridge;
 import org.openhab.core.thing.Thing;
+import org.openhab.core.thing.ThingRegistry;
 import org.openhab.core.thing.ThingTypeUID;
 import org.openhab.core.thing.binding.BaseThingHandlerFactory;
 import org.openhab.core.thing.binding.ThingHandler;
 import org.openhab.core.thing.binding.ThingHandlerFactory;
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Deactivate;
 import org.osgi.service.component.annotations.Reference;
 import org.smarthomej.binding.tuya.internal.handler.ProjectHandler;
 import org.smarthomej.binding.tuya.internal.handler.TuyaDeviceHandler;
-import org.smarthomej.binding.tuya.internal.local.UdpDiscoveryListener;
-import org.smarthomej.binding.tuya.internal.util.SchemaDp;
+import org.smarthomej.binding.tuya.internal.handler.TuyaGatewayHandler;
+import org.smarthomej.binding.tuya.internal.handler.TuyaSubDeviceHandler;
+import org.smarthomej.binding.tuya.internal.local.TuyaDeviceManagerFactory;
+import org.smarthomej.binding.tuya.internal.schema.SchemaRegistry;
 import org.smarthomej.commons.SimpleDynamicCommandDescriptionProvider;
 
 import com.google.gson.Gson;
-import com.google.gson.reflect.TypeToken;
-
-import io.netty.channel.EventLoopGroup;
-import io.netty.channel.nio.NioEventLoopGroup;
 
 /**
  * The {@link TuyaHandlerFactory} is responsible for creating things and thing
@@ -55,32 +51,27 @@ import io.netty.channel.nio.NioEventLoopGroup;
 @Component(configurationPid = "binding.tuya", service = ThingHandlerFactory.class)
 @SuppressWarnings("unused")
 public class TuyaHandlerFactory extends BaseThingHandlerFactory {
+
     private static final Set<ThingTypeUID> SUPPORTED_THING_TYPES_UIDS = Set.of(THING_TYPE_PROJECT,
-            THING_TYPE_TUYA_DEVICE);
-    private static final Type STORAGE_TYPE = TypeToken.getParameterized(List.class, SchemaDp.class).getType();
+            THING_TYPE_TUYA_GATEWAY, THING_TYPE_TUYA_DEVICE, THING_TYPE_TUYA_SUB_DEVICE);
 
     private final SimpleDynamicCommandDescriptionProvider dynamicCommandDescriptionProvider;
     private final HttpClient httpClient;
     private final Gson gson = new Gson();
-    private final UdpDiscoveryListener udpDiscoveryListener;
-    private final EventLoopGroup eventLoopGroup;
-    private final Storage<String> storage;
+    private final TuyaDeviceManagerFactory tuyaDeviceManagerFactory;
+    private final ThingRegistry thingRegistry;
+    private final SchemaRegistry schemaRegistry;
 
     @Activate
     public TuyaHandlerFactory(@Reference HttpClientFactory httpClientFactory,
             @Reference SimpleDynamicCommandDescriptionProvider dynamicCommandDescriptionProvider,
-            @Reference StorageService storageService) {
+            @Reference StorageService storageService, @Reference TuyaDeviceManagerFactory tuyaDeviceManagerFactory,
+            @Reference ThingRegistry thingRegistry, @Reference SchemaRegistry schemaRegistry) {
         this.httpClient = httpClientFactory.getCommonHttpClient();
         this.dynamicCommandDescriptionProvider = dynamicCommandDescriptionProvider;
-        this.eventLoopGroup = new NioEventLoopGroup();
-        this.udpDiscoveryListener = new UdpDiscoveryListener(eventLoopGroup);
-        this.storage = storageService.getStorage("org.smarthomej.binding.tuya.Schema");
-    }
-
-    @Deactivate
-    public void deactivate() {
-        udpDiscoveryListener.deactivate();
-        eventLoopGroup.shutdownGracefully();
+        this.tuyaDeviceManagerFactory = tuyaDeviceManagerFactory;
+        this.thingRegistry = thingRegistry;
+        this.schemaRegistry = schemaRegistry;
     }
 
     @Override
@@ -93,12 +84,16 @@ public class TuyaHandlerFactory extends BaseThingHandlerFactory {
         ThingTypeUID thingTypeUID = thing.getThingTypeUID();
 
         if (THING_TYPE_PROJECT.equals(thingTypeUID)) {
-            return new ProjectHandler(thing, httpClient, storage, gson);
+            return new ProjectHandler(thing, httpClient, schemaRegistry, gson, thingRegistry);
+        } else if (THING_TYPE_TUYA_GATEWAY.equals(thingTypeUID)) {
+            return new TuyaGatewayHandler((Bridge) thing, tuyaDeviceManagerFactory);
+        } else if (THING_TYPE_TUYA_SUB_DEVICE.equals(thingTypeUID)) {
+            return new TuyaSubDeviceHandler(thing, schemaRegistry.get(thing.getUID().getId()),
+                    dynamicCommandDescriptionProvider);
         } else if (THING_TYPE_TUYA_DEVICE.equals(thingTypeUID)) {
-            return new TuyaDeviceHandler(thing, gson.fromJson(storage.get(thing.getUID().getId()), STORAGE_TYPE), gson,
-                    dynamicCommandDescriptionProvider, eventLoopGroup, udpDiscoveryListener);
+            return new TuyaDeviceHandler(thing, schemaRegistry.get(thing.getUID().getId()),
+                    dynamicCommandDescriptionProvider, tuyaDeviceManagerFactory);
         }
-
         return null;
     }
 }
